tasks/prediction/requirements.txt
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudevents==1.10.1
db-dtypes==1.2.0
deprecation==2.1.0
fiona==1.9.6
Flask==3.0.3
functions-framework==3.5.0
geopandas==0.14.3
google-api-core==2.18.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-cloud-bigquery==3.20.1
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.4.0
MarkupSafe==2.1.5
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pandas==2.2.2
pandas-gbq==0.22.0
patsy==0.5.6
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydata-google-auth==1.8.2
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.4.2
scipy==1.13.0
shapely==2.0.3
six==1.16.0
statsmodels==0.14.1
threadpoolctl==3.4.0
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.0
Werkzeug==3.0.2
